Revert "Support TypeScript"
This reverts commit 1b1b6363d63a409672ae9af02f3f9e25b8ddc199.
