--- src/index.ts
@@ -1,43 +1,0 @@
-import chalk from 'chalk';
-import { marked } from 'marked';
-import TerminalRenderer from 'marked-terminal';
-import dedent from 'dedent';
-
-const defaultOptions = {
-  code: chalk.magenta,
-  blockquote: chalk.gray,
-  html: chalk.gray,
-  heading: chalk.green.bold,
-  firstHeading: chalk.magenta.underline.bold,
-  strong: chalk.bold,
-  em: chalk.italic,
-  codespan: chalk.blueBright,
-  del: chalk.dim.strikethrough,
-  link: chalk.cyan,
-  href: chalk.cyan.underline,
-  tab: 2,
-};
-
-marked.setOptions({
-  renderer: new TerminalRenderer(defaultOptions),
-});
-
-export default function log (message: string, gutter?: string) {
-  // Replace line breaks with \n and remove top-level quotes
-  // added by JSON.stringify so marked can interpret it as
-  // markdown and can pretty print to console
-  const unquotedMessage = JSON.stringify(message).replace(/"(.*)"/, '$1');
-
-  // Add marked to parse Markdown
-  const multiColorMessage = marked(unquotedMessage);
-
-  // Finally replace \n with real line-breaks
-  const prettyPrintMessage = multiColorMessage.replace(/\\n/g, '\n');
-
-  const dedentLines = dedent(prettyPrintMessage);
-
-  if (gutter) {
-    return console.log(`\n${dedentLines}\n`);
-  }
-  console.log(dedentLines);
-};
