--- test/module.test.ts
@@ -1,7 +1,0 @@
-import log from '../src';
-
-describe('blah', () => {
-  it('works', () => {
-    expect(log('## hello world')).toEqual('hello world');
-  });
-});
